Move the loader bootstrap into a JS file and let it apply visibility
The script that adapts the classic screen to the iframe now lives in
meta-box-iframe.js, printed inline by the PHP so that it is not a
registered script anything can depend on. It also applies the editor's
meta box visibility preferences itself: a style built from the parent's
preferences hides boxes before first paint, and the boxes follow later
changes through a parent store subscription, so the editor side needs
no code for it. The classic Screen Options hidden state no longer
applies inside the iframes, which matches how the block editor treats
it. The save flow finds the iframes by their frame names through the
elements; the browsing context names are cleared under the editor's
Document-Isolation-Policy, so window.frames cannot resolve them.
