+++ packages/edit-post/src/store/index.js
@@ -2,9 +2,7 @@ import { createReduxStore, register } from '@wordpress/data';
 import reducer from './reducer';
 import * as actions from './actions';
 import * as selectors from './selectors';
-import * as privateSelectors from './private-selectors';
 import { STORE_NAME } from './constants';
-import { unlock } from '../lock-unlock';
 
 /**
  * Store definition for the edit post namespace.
@@ -19,4 +17,3 @@ export const store = createReduxStore( STORE_NAME, {
 	selectors,
 } );
 register( store );
-unlock( store ).registerPrivateSelectors( privateSelectors );
